+++ component-classes/index.d.ts
@@ -495,8 +495,10 @@ export namespace backwardsCompatibleClasses {
     export { radioDisabled_1 as radioDisabled };
     export let modalTitle: string;
     export let pillClose: string;
+    export let removedAlertTokens: string;
     export let removedBadgeTokens: string;
     export let removedBoxTokens: string;
+    export let removedBreadcrumbsTokens: string;
 }
 export namespace pagination {
     export let currentPage: string;